+++ packages/javascript/lib/node/os/index.ts
@@ -3,4 +3,5 @@ export { getHostname } from './getHostname.js';
 export { getMachineId } from './getMachineId.js';
 export { getRamSize } from './getRamSize.js';
 export { getSid } from './getSid.js';
+export { getUptime } from './getUptime.js';
 export { runCommand } from './runCommand.js';