+++ packages/javascript/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.11.7 (2026--)
 
 - `fetchData`: Minor improvements
+- `getUptime`: Add `getUptime` method
 
 ## 1.11.6 (2026-04-12)